[FIX] runbot_gitlab_minimal: Each class into a one file
