+++ runbot_gitlab_minimal/models/__init__.py
@@ -3,3 +3,5 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from . import runbot_repo
+from . import runbot_build
+from . import runbot_branch
