+++ runbot_gitlab_minimal/models/runbot_repo.py
@@ -9,7 +9,7 @@ import urllib
 
 import requests
 
-from openerp import models, fields, api
+from openerp import models, fields
 
 _logger = logging.getLogger(__name__)
 
@@ -120,90 +120,3 @@ class RunbotRepo(models.Model):
                                       payload)
                 else:
                     raise
-
-
-class RunbotBuild(models.Model):
-    _inherit = "runbot.build"
-
-    def github_status(self, cr, uid, ids, context=None):
-        runbot_domain = self.pool['runbot.repo'].domain(cr, uid)
-        for build in self.browse(cr, uid, ids, context=context):
-            is_merge_request = build.branch_id.branch_name.isdigit()
-            source_project_id = False
-            _url = _get_url('/projects/:owner/:repo/statuses/%s' % build.name,
-                            build.repo_id.base)
-            if not build.repo_id.uses_gitlab:
-                super(RunbotBuild, self).github_status(cr, uid, ids,
-                                                       context=context)
-                continue
-            if not build.repo_id.token:
-                continue
-            session = _get_session(build.repo_id.token)
-            try:
-                if is_merge_request:
-                    url = _get_url('/projects/:owner/:repo/merge_requests/'
-                                   '?iid=%s' % build.branch_id.branch_name,
-                                   build.repo_id.base)
-                    response = session.get(url)
-                    response.raise_for_status()
-                    json = response.json()[0]
-                    source_project_id = json['source_project_id']
-                    if source_project_id:
-                        url = _get_url('/projects/%s' % source_project_id,
-                                       build.repo_id.base)
-                        response = session.get(url)
-                        response.raise_for_status()
-                        json = response.json()
-                        base_url = (json['web_url'].replace('http://', '').
-                                    replace('https://', ''))
-                        _url = _get_url('/projects/:owner/:repo/statuses/%s' %
-                                        build.name, base_url)
-                desc = "runbot build %s" % (build.dest,)
-                if build.state == 'testing':
-                    state = 'running'
-                elif build.state in ('running', 'done'):
-                    state = 'failed'
-                    if build.result == 'ok':
-                        state = 'success'
-                    if build.result == 'ko':
-                        state = 'failed'
-                    if build.result == 'skipped':
-                        state = 'canceled'
-                    if build.result == 'killed':
-                        state = 'canceled'
-                    desc += " (runtime %ss)" % (build.job_time,)
-                else:
-                    continue
-                status = {
-                    "state": state,
-                    "target_url": "http://%s/runbot/build/%s" % (runbot_domain,
-                                                                 build.id),
-                    "description": desc,
-                    "context": "ci/runbot"
-                }
-                _logger.debug("gitlab updating status %s to %s", build.name,
-                              state)
-                response = session.post(_url, status)
-                response.raise_for_status()
-            except Exception:
-                _logger.exception('gitlab API error %s', url)
-
-
-class RunbotBranch(models.Model):
-    _inherit = "runbot.branch"
-
-    branch_url = fields.Char(compute='_get_branch_url')
-
-    @api.multi
-    def _get_branch_url(self):
-        _branch_urls = super(RunbotBranch, self)._get_branch_url(None, None)
-        for branch in self:
-            if not branch.repo_id.uses_gitlab:
-                branch.branch_url = _branch_urls[branch.id]
-            else:
-                if branch.branch_name.isdigit():
-                    branch.branch_url = "https://%s/merge_requests/%s" % (
-                        branch.repo_id.base, branch.branch_name)
-                else:
-                    branch.branch_url = ("https://%s/tree/%s" % (
-                        branch.repo_id.base, branch.branch_name))
